updated flags for command testing registration
